refactor: 리팩토링 한 white, blue 버튼 컴포넌트로 수정(#63)
