+++ frontend/src/components/Issues/Menu/MenuTab.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import ButtonGroup from "components/common/Button/ButtonGroup";
-import AddButton from "components/common/Button/AddButton";
+import AddButton from "components/common/Button/BlueButtons";
 import { Link } from "react-router-dom";
 import { useSetRecoilState } from "recoil";
 import {
@@ -29,7 +29,7 @@ const MenuTab = () => {
 				isMainPage={true}
 			/>
 			<Link to="main/new">
-				<AddButton text={"이슈 작성"} />
+				<AddButton text={"이슈 작성"} icon={"plus"} size={"m"} />
 			</Link>
 		</Wrapper>
 	);
